pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
name = "saxomoose.scraper"
version = "0.0.1"
dependencies = [
    "beautifulsoup4",
    "python-dotenv",
    "lxml",
    "playwright ==1.30.0",
    "python-dotenv",
    "redis",
]

[tool.setuptools.packages]
find = {}  # Scan the project directory with the default parameters

[tool.setuptools]
py-modules = []

[tool.isort]
combine_as_imports = true
lines_between_sections = 0
profile = "black"